Use an exponential factor when zooming
This makes the zoom experience a lot smoother: the step size no longer
depends on how close you're zoomed in.

+++ src/T2DMap.h
@@ -56,7 +56,7 @@ public:
     Q_DISABLE_COPY(T2DMap)
     explicit T2DMap(QWidget* parent = nullptr);
     void paintMap();
-    void setMapZoom(int zoom);
+    void setMapZoom(qreal zoom);
     QColor getColor(int id);
     void init();
     void exportAreaImage(int);
@@ -86,7 +86,7 @@ public:
 
     TMap* mpMap;
     QPointer<Host> mpHost;
-    int xyzoom;
+    qreal xyzoom;
     int mRX;
     int mRY;
     QPoint mPHighlight;
